+++ lib/recording_studio_oauth/wordpress_relay.rb
@@ -1,0 +1,35 @@
+# frozen_string_literal: true
+
+module RecordingStudioOauth
+  module WordPressRelay
+    CONNECT_PATH = "/wordpress/connect"
+    CALLBACK_PATH = "/wordpress/callback"
+
+    module_function
+
+    def connect_path
+      "#{mount_path}#{CONNECT_PATH}"
+    end
+
+    def callback_path
+      "#{mount_path}#{CALLBACK_PATH}"
+    end
+
+    def connect_url(base_url:)
+      "#{normalize_base(base_url)}#{connect_path}"
+    end
+
+    def callback_url(base_url:)
+      "#{normalize_base(base_url)}#{callback_path}"
+    end
+
+    def mount_path
+      path = RecordingStudioOauth.configuration.engine_mount_path.to_s
+      path.start_with?("/") ? path.chomp("/") : "/#{path}".chomp("/")
+    end
+
+    def normalize_base(base_url)
+      base_url.to_s.chomp("/")
+    end
+  end
+end